+++ elmyra.ip.access.epo/CHANGES.rst
@@ -13,6 +13,7 @@ development
 - api/cql: automatically apply number normalization to "num" fields, too
 - ui: query builder II: full cql field chooser
 - ui: perform query when hitting hotkey "meta+return" in query form field
+- ui: clipboard modifier intercepts when pasting text into empty query form field
 
 0.4.2
 =====